Added evaluation results logic
